@@ -1,13 +1,11 @@
 """Basic smoke test to validate integration test setup."""
 
-from typing import Optional
-
 import pytest
 
 from nutrient_dws import NutrientClient
 
 # Type annotation for mypy
-API_KEY: Optional[str] = None
+API_KEY: str | None = None
 
 try:
     from . import integration_config  # type: ignore[attr-defined]
