@@ -2,7 +2,6 @@
 
 import os
 from pathlib import Path
-from typing import Optional, Union
 
 import pytest
 
@@ -11,8 +10,8 @@ from nutrient_dws import NutrientClient
 try:
     from . import integration_config  # type: ignore[attr-defined]
 
-    API_KEY: Optional[str] = integration_config.API_KEY
-    BASE_URL: Optional[str] = getattr(integration_config, "BASE_URL", None)
+    API_KEY: str | None = integration_config.API_KEY
+    BASE_URL: str | None = getattr(integration_config, "BASE_URL", None)
     TIMEOUT: int = getattr(integration_config, "TIMEOUT", 60)
 except ImportError:
     API_KEY = None
@@ -20,7 +19,7 @@ except ImportError:
     TIMEOUT = 60
 
 
-def assert_is_pdf(file_path_or_bytes: Union[str, bytes]) -> None:
+def assert_is_pdf(file_path_or_bytes: str | bytes) -> None:
     """Assert that a file or bytes is a valid PDF."""
     if isinstance(file_path_or_bytes, str):
         with open(file_path_or_bytes, "rb") as f:
